cln: Extract gridsearch functionality (mostly) to subpackage
