+++ src/gen_experiments/__init__.py
@@ -5,9 +5,10 @@ from typing import Any
 
 import numpy as np
 from numpy.typing import NDArray
-from pysindy import BaseDifferentiation, FiniteDifference, SINDy  # type: ignore
+from pysindy import BaseDifferentiation, FiniteDifference, SINDy
 
-from . import gridsearch, odes, pdes
+from . import gridsearch  # type: ignore
+from . import odes, pdes
 from .utils import SINDyTrialData, make_model  # noqa: F401
 
 this_module = importlib.import_module(__name__)
