+++ src/gen_experiments/gridsearch/__init__.py
@@ -1,5 +1,5 @@
+from collections.abc import Iterable
 from copy import copy
-from dataclasses import dataclass, field
 from functools import partial
 from logging import getLogger
 from pprint import pformat
@@ -9,14 +9,15 @@ from typing import (
     Any,
     Callable,
     Collection,
-    Iterable,
     Optional,
     Sequence,
     TypeVar,
+    cast,
 )
 
 import matplotlib.pyplot as plt
 import numpy as np
+from matplotlib.axes import Axes
 from numpy.typing import DTypeLike, NDArray
 from scipy.stats import kstest
 
@@ -24,18 +25,17 @@ import gen_experiments
 from gen_experiments import config
 from gen_experiments.odes import plot_ode_panel
 from gen_experiments.plotting import _PlotPrefs
-from gen_experiments.utils import (
+from gen_experiments.typing import FloatND
+from gen_experiments.utils import simulate_test_data
+
+from .typing import (
+    ExpResult,
     GridsearchResult,
     GridsearchResultDetails,
     NestedDict,
-    SavedData,
+    SavedGridPoint,
     SeriesDef,
     SeriesList,
-    SINDyTrialData,
-    _amax_to_full_inds,
-    _argopt,
-    _grid_locator_match,
-    simulate_test_data,
 )
 
 pformat = partial(pformat, indent=4, sort_dicts=True)
@@ -47,30 +47,99 @@ OtherSliceDef = tuple[int | Callable]
 SkinnySpecs = Optional[tuple[tuple[str, ...], tuple[OtherSliceDef, ...]]]
 
 
-@dataclass(frozen=True)
-class GridLocator:
-    """A specification of which points in a gridsearch to match.
-
-    Rather than specifying the exact point in the mega-grid of every
-    varied axis, specify by result, e.g "all of the points from the
-    Kalman series that had the best mean squared error as noise was
-    varied.
+def _amax_to_full_inds(
+    amax_inds: Collection[tuple[int | slice, int] | ellipsis] | ellipsis,
+    amax_arrays: list[list[GridsearchResult[np.void]]],
+) -> set[tuple[int, ...]]:
+    """Find full indexers to selected elements of argmax arrays
 
     Args:
-        metric: The metric in which to find results.  An ellipsis means "any metrics"
-        keep_axis: The grid-varied parameter in which to find results, or a tuple of
-            that axis and position along that axis.  To search a particular value of
-            that parameter, use the param_match kwarg.  An ellipsis means "any axis"
-        param_match: A collection of dictionaries to match parameter values represented
-            by points in the gridsearch.  Dictionary equality is checked for every
-            non-callable value; for callable values, it is applied to the grid
-            parameters and must return a boolean.  Logical OR is applied across the
-            collection
+        amax_inds: selection statemtent of which argmaxes to return.
+        amax_arrays: arrays of indexes to full gridsearch that are responsible for
+            the computed max values.  First level of nesting reflects series, second
+            level reflects which grid grid axis.
+    Returns:
+        all indexers to full gridsearch that are requested by amax_inds
     """
 
-    metric: str | ellipsis = field(default=...)
-    keep_axis: str | tuple[str, int] | ellipsis = field(default=...)
-    param_match: Collection[dict[str, Any]] = field(default=())
+    def np_to_primitive(tuple_like: np.void) -> tuple[int, ...]:
+        return tuple(int(el) for el in cast(Iterable, tuple_like))
+
+    if amax_inds is ...:  # grab each element from arrays in list of lists of arrays
+        return {
+            np_to_primitive(el)
+            for ar_list in amax_arrays
+            for arr in ar_list
+            for el in arr.flatten()
+        }
+    all_inds = set()
+    for plot_axis_results in [el for series in amax_arrays for el in series]:
+        for ind in amax_inds:
+            if ind is ...:  # grab each element from arrays in list of lists of arrays
+                all_inds |= {
+                    np_to_primitive(el)
+                    for ar_list in amax_arrays
+                    for arr in ar_list
+                    for el in arr.flatten()
+                }
+            elif isinstance(ind[0], int):
+                all_inds |= {np_to_primitive(cast(np.void, plot_axis_results[ind]))}
+            else:  # ind[0] is slice(None)
+                all_inds |= {np_to_primitive(el) for el in plot_axis_results[ind]}
+    return all_inds
+
+
+_EqTester = TypeVar("_EqTester")
+
+
+def _param_normalize(val: _EqTester) -> _EqTester | str:
+    if type(val).__eq__ == object.__eq__:
+        return repr(val)
+    else:
+        return val
+
+
+def _grid_locator_match(
+    exp_params: dict[str, Any],
+    exp_ind: tuple[int, ...],
+    param_spec: Collection[dict[str, Any]],
+    ind_spec: Collection[tuple[int, ...]],
+) -> bool:
+    """Determine whether experimental parameters match a specification
+
+    Logical clause applied is:
+
+        OR((exp_params MATCHES params for params in param_spec))
+        AND
+        OR((exp_ind MATCHES ind for ind in ind_spec))
+
+    Treats OR of an empty collection as falsy
+    Args:
+        exp_params: the experiment parameters to evaluate
+        exp_ind: the experiemnt's full-size grid index to evaluate
+        param_spec: the criteria for matching exp_params
+        ind_spec: the criteria for matching exp_ind
+    """
+    found_match = False
+    for params_or in param_spec:
+        params_or = {k: _param_normalize(v) for k, v in params_or.items()}
+
+        try:
+            if all(
+                _param_normalize(exp_params[param]) == value
+                for param, value in params_or.items()
+            ):
+                found_match = True
+                break
+        except KeyError:
+            pass
+    for ind_or in ind_spec:
+        # exp_ind doesn't include metric, so skip first metric
+        if _index_in(exp_ind, ind_or[1:]):
+            break
+    else:
+        return False
+    return found_match
 
 
 def run(
@@ -78,7 +147,7 @@ def run(
     group: str,
     grid_params: list[str],
     grid_vals: list[Sequence],
-    grid_decisions: Sequence[str],
+    grid_decisions: list[str],
     other_params: dict,
     series_params: Optional[SeriesList] = None,
     metrics: Sequence[str] = (),
@@ -118,8 +187,8 @@ def run(
     metric_ordering = [base_ex.metric_ordering[metric] for metric in metrics]
     n_plotparams = len([decide for decide in grid_decisions if decide == "plot"])
     series_searches: list[tuple[list[GridsearchResult], list[GridsearchResult]]] = []
-    intermediate_data: list[SavedData] = []
-    plot_data: list[SavedData] = []
+    intermediate_data: list[SavedGridPoint] = []
+    plot_data: list[SavedGridPoint] = []
     if base_group is not None:
         other_params["group"] = base_group
     for s_counter, series_data in enumerate(series_params.series_list):
@@ -145,14 +214,12 @@ def run(
         for ind_counter, ind in enumerate(gridpoint_selector):
             print(f"Calculating series {s_counter}, gridpoint{ind_counter}", end="\r")
             new_seed = rng.integers(1000)
-            param_updates = {}
             for axis_ind, key, val_list in zip(ind, new_grid_params, new_grid_vals):
-                param_updates[key] = val_list[axis_ind]
-                curr_other_params.update(param_updates)
+                curr_other_params[key] = val_list[axis_ind]
             curr_results, grid_data = base_ex.run(
                 new_seed, **curr_other_params, display=False, return_all=True
             )
-            grid_data: SINDyTrialData
+            grid_data: ExpResult
             intermediate_data.append(
                 {"params": curr_other_params.flatten(), "pind": ind, "data": grid_data}
             )
@@ -205,7 +272,6 @@ def run(
             series_searches, (ser.name for ser in series_params.series_list)
         ):
             plot(
-                fig,
                 subplots,
                 metrics,
                 grid_params,
@@ -244,8 +310,7 @@ def run(
 
 
 def plot(
-    fig: plt.Figure,
-    subplots: Sequence[plt.Axes],
+    subplots: NDArray[Annotated[np.void, "Axes"]],
     metrics: Sequence[str],
     grid_params: Sequence[str],
     grid_vals: Sequence[Sequence[float] | np.ndarray],
@@ -253,11 +318,13 @@ def plot(
     name: str,
     legends: bool,
 ):
+    if len(metrics) == 0:
+        raise ValueError("Nothing to plot")
     for m_ind_row, m_name in enumerate(metrics):
         for col, (param_name, x_ticks, param_search) in enumerate(
             zip(grid_params, grid_vals, grid_searches)
         ):
-            ax = subplots[m_ind_row, col]
+            ax = cast(Axes, subplots[m_ind_row, col])
             ax.plot(x_ticks, param_search[m_ind_row], label=name)
             x_ticks = np.array(x_ticks)
             if m_name in ("coeff_mse", "coeff_mae"):
@@ -278,16 +345,54 @@ def plot(
             if col == 0:
                 ax.set_ylabel(f"{m_name}")
     if legends:
-        ax.legend()
+        ax.legend()  # type: ignore
 
 
 T = TypeVar("T", bound=np.generic)
 
 
+def _argopt(
+    arr: FloatND, axis: Optional[int | tuple[int, ...]] = None, opt: str = "max"
+) -> NDArray[np.void]:
+    """Calculate the argmax/min, but accept tuple axis.
+
+    Ignores NaN values
+
+    Args:
+        arr: an array to search
+        axis: The axis or axes to search through for the argmax/argmin.
+        opt: One of {"max", "min"}
+
+    Returns:
+        array of indices for the argopt.  If m = arr.ndim and n = len(axis),
+        the final result will be an array of ndim = m-n with elements being
+        tuples of length m
+    """
+    dtype: DTypeLike = [(f"f{axind}", "i") for axind in range(arr.ndim)]
+    if axis is None:
+        axis = ()
+    axis = (axis,) if isinstance(axis, int) else axis
+    keep_axes = tuple(sorted(set(range(arr.ndim)) - set(axis)))
+    keep_shape = tuple(arr.shape[ax] for ax in keep_axes)
+    result = np.empty(keep_shape, dtype=dtype)
+    optfun = np.nanargmax if opt == "max" else np.nanargmin
+    for slise in np.ndindex(keep_shape):
+        sub_arr = arr
+        # since we shrink shape, we need to chop of axes from the end
+        for ind, ax in zip(reversed(slise), reversed(keep_axes)):
+            sub_arr = np.take(sub_arr, ind, ax)
+        subind_max = np.unravel_index(optfun(sub_arr), sub_arr.shape)
+        fullind_max = np.empty((arr.ndim), int)
+        fullind_max[np.array(keep_axes, int)] = slise
+        fullind_max[np.array(axis, int)] = subind_max
+        result[slise] = tuple(fullind_max)
+    return result
+
+
 def _marginalize_grid_views(
     grid_decisions: Iterable[str],
     results: Annotated[NDArray[T], "shape (n_metrics, *n_gridsearch_values)"],
-    max_or_min: Sequence[str] = None,
+    max_or_min: Sequence[str],
 ) -> tuple[list[GridsearchResult[T]], list[GridsearchResult]]:
     """Marginalize unnecessary dimensions by taking max across axes.
 
@@ -303,7 +408,7 @@ def _marginalize_grid_views(
         a list of the metric optima for each plottable grid decision, and
         a list of the flattened argoptima.
     """
-    arg_dtype: DTypeLike = ",".join(results.ndim * "i")
+    arg_dtype = np.dtype(",".join(results.ndim * "i"))
     plot_param_inds = [ind for ind, val in enumerate(grid_decisions) if val == "plot"]
     grid_searches = []
     args_maxes = []
@@ -315,7 +420,11 @@ def _marginalize_grid_views(
         )
         sub_arrs = []
         for m_ind, (result, opt) in enumerate(zip(results, max_or_min)):
-            pad_m_ind = np.vectorize(lambda tp: np.void((m_ind, *tp), dtype=arg_dtype))
+
+            def _metric_pad(tp: tuple[int, ...]) -> np.void:
+                return np.void((m_ind, *tp), dtype=arg_dtype)
+
+            pad_m_ind = np.vectorize(_metric_pad)
             arg_max = pad_m_ind(_argopt(result, reduce_axes, opt))
             sub_arrs.append(arg_max)
 
@@ -326,7 +435,7 @@ def _marginalize_grid_views(
 
 
 def _ndindex_skinny(
-    shape: tuple[int],
+    shape: tuple[int, ...],
     thin_axes: Optional[Sequence[int]] = None,
     thin_slices: Optional[Sequence[OtherSliceDef]] = None,
 ):
@@ -360,6 +469,7 @@ def _ndindex_skinny(
         n_thin = len(thin_axes)
         thin_slices = n_thin * ((n_thin - 1) * (0,),)
     full_indexes = np.ndindex(shape)
+    thin_slices = cast(Sequence[OtherSliceDef], thin_slices)
 
     def ind_checker(multi_index):
         """Check if a multi_index meets thin index criteria"""
@@ -415,3 +525,59 @@ def _curr_skinny_specs(
         )
         where_others.append(new_criteria)
     return skinny_param_inds, tuple(where_others)
+
+
+def strict_find_grid_match(
+    results: GridsearchResultDetails,
+    *,
+    params: Optional[dict[str, Any]] = None,
+    ind_spec: Optional[tuple[int | slice, int] | ellipsis] = None,
+) -> ExpResult:
+    if params is None:
+        params = {}
+    if ind_spec is None:
+        ind_spec = ...
+    matches = []
+    amax_arrays = [
+        [single_ser_and_axis[1] for single_ser_and_axis in single_series_all_axes]
+        for _, single_series_all_axes in results["series_data"].items()
+    ]
+    full_inds = _amax_to_full_inds((ind_spec,), amax_arrays)
+
+    for trajectory in results["plot_data"]:
+        if _grid_locator_match(
+            trajectory["params"], trajectory["pind"], (params,), full_inds
+        ):
+            matches.append(trajectory)
+
+    if len(matches) > 1:
+        raise ValueError("Specification is nonunique; matched multiple results")
+    if len(matches) == 0:
+        raise ValueError("Could not find a match")
+    return matches[0]["data"]
+
+
+def _index_in(base: tuple[int, ...], tgt: tuple[int | ellipsis | slice, ...]) -> bool:
+    """Determine whether base indexing tuple will match given numpy index"""
+    if len(base) > len(tgt):
+        return False
+    curr_ax = 0
+    for ax, ind in enumerate(tgt):
+        if isinstance(ind, int):
+            try:
+                if ind != base[curr_ax]:
+                    return False
+            except IndexError:
+                return False
+        elif isinstance(ind, slice):
+            if not (ind.start is None and ind.stop is None and ind.step is None):
+                raise ValueError("Only slices allowed are `slice(None)`")
+        elif ind is ...:
+            base_ind_remaining = len(base) - curr_ax
+            tgt_ind_remaining = len(tgt) - ax
+            # ellipsis can take 0 or more spots
+            curr_ax += max(base_ind_remaining - tgt_ind_remaining, -1)
+        curr_ax += 1
+    if curr_ax == len(base):
+        return True
+    return False
