+++ tests/test_all.py
@@ -1,7 +1,8 @@
 import numpy as np
 import pytest
 
-from gen_experiments import gridsearch, utils
+import gen_experiments.gridsearch.typing
+from gen_experiments import gridsearch
 
 
 def test_thin_indexing():
@@ -82,9 +83,9 @@ def test_marginalize_grid_views():
 
 
 def test_argopt_tuple_axis():
-    arr = np.arange(16).reshape(2, 2, 2, 2)
+    arr = np.arange(16, dtype=np.float_).reshape(2, 2, 2, 2)
     arr[0, 0, 0, 0] = 1000
-    result = utils._argopt(arr, (1, 3))
+    result = gridsearch._argopt(arr, (1, 3))
     expected = np.array(
         [[(0, 0, 0, 0), (0, 1, 1, 1)], [(1, 1, 0, 1), (1, 1, 1, 1)]], dtype="i,i,i,i"
     )
@@ -92,18 +93,18 @@ def test_argopt_tuple_axis():
 
 
 def test_argopt_empty_tuple_axis():
-    arr = np.arange(4).reshape(4)
-    result = utils._argopt(arr, ())
+    arr = np.arange(4, dtype=np.float_).reshape(4)
+    result = gridsearch._argopt(arr, ())
     expected = np.array([(0,), (1,), (2,), (3,)], dtype=[("f0", "i")])
     np.testing.assert_array_equal(result, expected)
-    result = utils._argopt(arr, None)
+    result = gridsearch._argopt(arr, None)
     pass
 
 
 def test_argopt_int_axis():
-    arr = np.arange(8).reshape(2, 2, 2)
+    arr = np.arange(8, dtype=np.float_).reshape(2, 2, 2)
     arr[0, 0, 0] = 1000
-    result = utils._argopt(arr, 1)
+    result = gridsearch._argopt(arr, 1)
     expected = np.array([[(0, 0, 0), (0, 1, 1)], [(1, 1, 0), (1, 1, 1)]], dtype="i,i,i")
     np.testing.assert_array_equal(result, expected)
 
@@ -112,19 +113,21 @@ def test_index_in():
     match_me = (1, ..., slice(None), 3)
     good = [(1, 2, 1, 3), (1, 1, 3)]
     for g in good:
-        assert utils._index_in(g, match_me)
+        assert gridsearch._index_in(g, match_me)
     bad = [(1, 3), (1, 1, 2), (1, 1, 1, 2)]
     for b in bad:
-        assert not utils._index_in(b, match_me)
+        assert not gridsearch._index_in(b, match_me)
 
 
 def test_index_in_errors():
     with pytest.raises(ValueError):
-        utils._index_in((1,), (slice(-1),))
+        gridsearch._index_in((1,), (slice(-1),))
 
 
 def test_flatten_nested_dict():
-    deep = utils.NestedDict(a=utils.NestedDict(b=1))
+    deep = gen_experiments.gridsearch.typing.NestedDict(
+        a=gen_experiments.gridsearch.typing.NestedDict(b=1)
+    )
     result = deep.flatten()
     assert deep != result
     expected = {"a.b": 1}
@@ -154,7 +157,7 @@ def test_grid_locator_match():
         ),
     ]
     for param_spec, ind_spec in good_specs:
-        assert utils._grid_locator_match(m_params, m_ind, param_spec, ind_spec)
+        assert gridsearch._grid_locator_match(m_params, m_ind, param_spec, ind_spec)
 
     bad_specs = [
         ((), ((0, 1),)),
@@ -164,7 +167,7 @@ def test_grid_locator_match():
         ((), ()),
     ]
     for param_spec, ind_spec in bad_specs:
-        assert not utils._grid_locator_match(m_params, m_ind, param_spec, ind_spec)
+        assert not gridsearch._grid_locator_match(m_params, m_ind, param_spec, ind_spec)
 
 
 def test_amax_to_full_inds():
@@ -181,7 +184,9 @@ def test_amax_to_full_inds():
 
 def test_flatten_nested_bad_dict():
     with pytest.raises(TypeError, match="keywords must be strings"):
-        utils.NestedDict(**{1: utils.NestedDict(b=1)})
+        gen_experiments.gridsearch.typing.NestedDict(
+            **{1: gen_experiments.gridsearch.typing.NestedDict(b=1)}
+        )
     with pytest.raises(TypeError, match="Only string keys allowed"):
-        deep = utils.NestedDict(a={1: 1})
+        deep = gen_experiments.gridsearch.typing.NestedDict(a={1: 1})
         deep.flatten()
